[CARBONDATA-2710][Spark Integration] Refactor CarbonSparkSqlParser for better code reuse.
Refactor CarbonSparkSqlParser for better code reuse

This closes #2466
